Split flatten into a pass
