+++ test/python/pulse/test_pulse_ir.py
@@ -36,22 +36,6 @@ from qiskit.pulse.compiler import MapMixedFrame, SetSequence, SetSchedule
 
 class TestSequenceIR(QiskitTestCase):
     """Test SequenceIR objects"""
-
-    def _get_scheduling_pm(self) -> PulseIrTranspiler:
-        pm = PulseIrTranspiler()
-        pm.append(MapMixedFrame())
-        pm.append(SetSequence())
-        pm.append(SetSchedule())
-        return pm
-
-    def _compare_scheduled_elements(self, list1, list2):
-        if len(list1) != len(list2):
-            return False
-        for x in list1:
-            if x not in list2:
-                return False
-        return True
-
     def test_ir_creation(self):
         """Test ir creation"""
         ir_example = SequenceIR(AlignLeft())
@@ -277,192 +261,6 @@ class TestSequenceIR(QiskitTestCase):
 
         with self.assertRaises(PulseError):
             ir_example.scheduled_elements(recursive=True)
-
-    def test_flatten_ir_no_sub_blocks(self):
-        """Test that flattening ir with no sub blocks doesn't do anything"""
-        ir_example = SequenceIR(AlignLeft())
-        inst = Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1))
-        inst2 = Play(Constant(100, 0.5), frame=QubitFrame(2), target=Qubit(2))
-        ir_example.append(inst)
-        ir_example.append(inst2)
-
-        ir_example = self._get_scheduling_pm().run(ir_example)
-
-        self.assertEqual(ir_example.flatten(), ir_example)
-
-    def test_flatten_inplace_flag(self):
-        """Test that inplace flag in flattening works"""
-        ir_example = SequenceIR(AlignLeft())
-        inst = Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1))
-        inst2 = Play(Constant(100, 0.5), frame=QubitFrame(2), target=Qubit(2))
-        ir_example.append(inst)
-        ir_example.append(inst2)
-
-        ir_example = self._get_scheduling_pm().run(ir_example)
-
-        self.assertTrue(ir_example.flatten(inplace=True) is ir_example)
-        self.assertFalse(ir_example.flatten() is ir_example)
-
-    def test_flatten_one_sub_block(self):
-        """Test that flattening works with one block"""
-        ir_example = SequenceIR(AlignLeft())
-        inst = Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1))
-        inst2 = Play(Constant(100, 0.5), frame=QubitFrame(2), target=Qubit(2))
-        block = SequenceIR(AlignLeft())
-        block.append(inst)
-        block.append(inst2)
-        ir_example.append(block)
-
-        ir_example = self._get_scheduling_pm().run(ir_example)
-
-        flat = ir_example.flatten()
-        edge_list = flat.sequence.edge_list()
-        print(edge_list)
-        self.assertEqual(len(edge_list), 4)
-        self.assertTrue((0, 5) in edge_list)
-        self.assertTrue((0, 6) in edge_list)
-        self.assertTrue((5, 1) in edge_list)
-        self.assertTrue((6, 1) in edge_list)
-        self.assertTrue((0, inst) in flat.scheduled_elements())
-        self.assertTrue((0, inst2) in flat.scheduled_elements())
-
-    def test_flatten_one_sub_block_and_parallel_instruction(self):
-        """Test that flattening works with one block and parallel instruction"""
-        ir_example = SequenceIR(AlignLeft())
-        inst = Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1))
-        inst2 = Play(Constant(100, 0.5), frame=QubitFrame(2), target=Qubit(2))
-        block = SequenceIR(AlignLeft())
-        block.append(inst)
-        block.append(inst2)
-        ir_example.append(block)
-        ir_example.append(Play(Constant(100, 0.5), frame=QubitFrame(3), target=Qubit(3)))
-
-        ir_example = self._get_scheduling_pm().run(ir_example)
-
-        flat = ir_example.flatten()
-        edge_list = flat.sequence.edge_list()
-        self.assertEqual(len(edge_list), 6)
-        self.assertTrue((0, 3) in edge_list)
-        self.assertTrue((3, 1) in edge_list)
-        self.assertTrue((0, 6) in edge_list)
-        self.assertTrue((0, 7) in edge_list)
-        self.assertTrue((6, 1) in edge_list)
-        self.assertTrue((7, 1) in edge_list)
-        self.assertTrue(
-            self._compare_scheduled_elements(
-                ir_example.scheduled_elements(recursive=True), flat.scheduled_elements()
-            )
-        )
-
-    def test_flatten_one_sub_block_and_sequential_instructions(self):
-        """Test that flattening works with one block and sequential instructions"""
-        ir_example = SequenceIR(AlignLeft())
-        inst = Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1))
-        inst2 = Play(Constant(100, 0.5), frame=QubitFrame(2), target=Qubit(2))
-        block = SequenceIR(AlignLeft())
-        block.append(inst)
-        block.append(inst2)
-        ir_example.append(Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1)))
-        ir_example.append(block)
-        ir_example.append(Play(Constant(100, 0.5), frame=QubitFrame(2), target=Qubit(2)))
-
-        ir_example = self._get_scheduling_pm().run(ir_example)
-
-        flat = ir_example.flatten()
-        edge_list = flat.sequence.edge_list()
-        self.assertEqual(len(edge_list), 8)
-        self.assertTrue((0, 2) in edge_list)
-        self.assertTrue((4, 1) in edge_list)
-        self.assertTrue((2, 7) in edge_list)
-        self.assertTrue((2, 8) in edge_list)
-        self.assertTrue((7, 1) in edge_list)
-        self.assertTrue((8, 1) in edge_list)
-        self.assertTrue((7, 4) in edge_list)
-        self.assertTrue((8, 4) in edge_list)
-        self.assertTrue(
-            self._compare_scheduled_elements(
-                ir_example.scheduled_elements(recursive=True), flat.scheduled_elements()
-            )
-        )
-
-    def test_flatten_two_sub_blocks(self):
-        """Test that flattening works with two sub blocks"""
-        inst1 = Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1))
-        inst2 = Play(Constant(200, 0.5), frame=QubitFrame(1), target=Qubit(1))
-
-        block1 = SequenceIR(AlignLeft())
-        block1.append(inst1)
-        block2 = SequenceIR(AlignLeft())
-        block2.append(inst2)
-
-        ir_example = SequenceIR(AlignLeft())
-        ir_example.append(block1)
-        ir_example.append(block2)
-
-        ir_example = self._get_scheduling_pm().run(ir_example)
-
-        flat = ir_example.flatten()
-        ref = SequenceIR(AlignLeft())
-        ref.append(inst1)
-        ref.append(inst2)
-
-        ref = self._get_scheduling_pm().run(ref)
-
-        self.assertEqual(flat, ref)
-        self.assertTrue(
-            self._compare_scheduled_elements(
-                ir_example.scheduled_elements(recursive=True), flat.scheduled_elements()
-            )
-        )
-
-    def test_flatten_two_levels(self):
-        """Test that flattening works with one block and sequential instructions"""
-        inst = Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1))
-
-        block1 = SequenceIR(AlignLeft())
-        block1.append(inst)
-        block = SequenceIR(AlignLeft())
-        block.append(inst)
-        block.append(block1)
-
-        ir_example = SequenceIR(AlignLeft())
-        ir_example.append(inst)
-        ir_example.append(block)
-
-        ir_example = self._get_scheduling_pm().run(ir_example)
-
-        flat = ir_example.flatten()
-        edge_list = flat.sequence.edge_list()
-        self.assertEqual(len(edge_list), 4)
-        self.assertTrue((0, 2) in edge_list)
-        self.assertTrue((2, 6) in edge_list)
-        self.assertTrue((6, 7) in edge_list)
-        self.assertTrue((7, 1) in edge_list)
-        self.assertEqual(flat.scheduled_elements()[0], (0, inst))
-        self.assertEqual(flat.scheduled_elements()[1], (100, inst))
-        self.assertEqual(flat.scheduled_elements()[2], (200, inst))
-        # Verify that nodes removed from the graph are also removed from _time_table.
-        self.assertEqual(
-            {x for x in flat.sequence.node_indices() if x not in (0, 1)}, flat._time_table.keys()
-        )
-        self.assertTrue(
-            self._compare_scheduled_elements(
-                ir_example.scheduled_elements(recursive=True), flat.scheduled_elements()
-            )
-        )
-
-    def test_flatten_not_scheduled(self):
-        """Test that flattening an unscheduled IR raises an error"""
-        ir_example = SequenceIR(AlignLeft())
-        inst = Play(Constant(100, 0.5), frame=QubitFrame(1), target=Qubit(1))
-        inst2 = Play(Constant(100, 0.5), frame=QubitFrame(2), target=Qubit(2))
-        block = SequenceIR(AlignLeft())
-        block.append(inst)
-        block.append(inst2)
-        ir_example.append(block)
-
-        with self.assertRaises(PulseError):
-            ir_example.flatten()
 
     def test_ir_equating_different_alignment(self):
         """Test equating of blocks with different alignment"""
